perf(intraday): specialize column codecs for time-series compression
The IntradayBar ClickHouse table now applies per-column codec stacks:
DoubleDelta + ZSTD on event_date (regular 60s spacing collapses to a
few bits per row), Gorilla + ZSTD on OHLC and volume (XOR-of-prior
encoding for slow-changing floats), and ZSTD on the FixedString FIGI
column.

volume becomes Float64 to match Massive's flat-file schema (which
records fractional shares for thin instruments) and unify the column
family under a single Gorilla+ZSTD codec.

Existing tables are not migrated; new tables created after upgrade
adopt the codecs automatically.

+++ data/intraday.go
@@ -57,7 +57,7 @@ func SaveIntradayBarsBatch(ctx context.Context, tbl string, conn driver.Conn, ba
 			bar.High,
 			bar.Low,
 			bar.Close,
-			uint64(bar.Volume),
+			bar.Volume,
 		); err != nil {
 			return fmt.Errorf("append intraday row: %w", err)
 		}
